Store the authorization token in persistent storage to aid restarts
When restarting a lot (for testing mostly), it occurred to me that the
envoy was spitting out 503 errors. These are a indication of "Too many
installer sessions active".
To fix this, we can just use an existing session, thus storing the
authentication token on persistent storage would fix this problem.