First token attempt should always be the envoy-procedure
Otherwise we might end up with a owner token. This will prevent polling
the installer endpoint urls, which in its turn will prevent the privilege
escalation.

So now we always query envoy procedure to fetch token, unless installer
account is disabled (which is set if envoy procedure does not return a
valid result too.